+++ vector/src/main/java/im/vector/notifications/NotificationUtils.kt
@@ -30,6 +30,7 @@ import android.net.Uri
 import android.os.Build
 import android.support.annotation.StringRes
 import android.support.v4.app.*
+import android.support.v4.content.ContextCompat
 import android.text.TextUtils
 import im.vector.BuildConfig
 import im.vector.R
@@ -38,7 +39,6 @@ import im.vector.activity.LockScreenActivity
 import im.vector.activity.VectorHomeActivity
 import im.vector.activity.VectorRoomActivity
 import im.vector.receiver.NotificationBroadcastReceiver
-import im.vector.ui.themes.ThemeUtils
 import im.vector.util.PreferencesManager
 import im.vector.util.startNotificationChannelSettingsIntent
 import org.matrix.androidsdk.util.Log
@@ -69,13 +69,14 @@ object NotificationUtils {
      * IDs for actions
      * ========================================================================================== */
 
-    private const val JOIN_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.JOIN_ACTION"
-    private const val REJECT_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.REJECT_ACTION"
-    private const val QUICK_LAUNCH_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.QUICK_LAUNCH_ACTION"
-    const val SMART_REPLY_ACTION_PREFIX = "${BuildConfig.APPLICATION_ID}.NotificationUtils.SMART_REPLY_ACTION_"
-    const val DISMISS_SUMMARY_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.DISMISS_SUMMARY_ACTION"
-    const val DISMISS_ROOM_NOTIF_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.DISMISS_ROOM_NOTIF_ACTION"
-    private const val TAP_TO_VIEW_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationUtils.TAP_TO_VIEW_ACTION"
+    private const val JOIN_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.JOIN_ACTION"
+    private const val REJECT_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.REJECT_ACTION"
+    private const val QUICK_LAUNCH_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.QUICK_LAUNCH_ACTION"
+    const val MARK_ROOM_READ_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.MARK_ROOM_READ_ACTION"
+    const val SMART_REPLY_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.SMART_REPLY_ACTION"
+    const val DISMISS_SUMMARY_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.DISMISS_SUMMARY_ACTION"
+    const val DISMISS_ROOM_NOTIF_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.DISMISS_ROOM_NOTIF_ACTION"
+    private const val TAP_TO_VIEW_ACTION = "${BuildConfig.APPLICATION_ID}.NotificationActions.TAP_TO_VIEW_ACTION"
 
     /* ==========================================================================================
      * IDs for channels
@@ -128,41 +129,45 @@ object NotificationUtils {
          * Default notification importance: shows everywhere, makes noise, but does not visually
          * intrude.
          */
-        val noisyChannel = NotificationChannel(NOISY_NOTIFICATION_CHANNEL_ID,
+        notificationManager.createNotificationChannel(NotificationChannel(NOISY_NOTIFICATION_CHANNEL_ID,
                 context.getString(R.string.notification_noisy_notifications),
                 NotificationManager.IMPORTANCE_DEFAULT)
-        noisyChannel.description = context.getString(R.string.notification_noisy_notifications)
-        noisyChannel.enableVibration(true)
-        noisyChannel.enableLights(true)
-        notificationManager.createNotificationChannel(noisyChannel)
+                .apply {
+                    description = context.getString(R.string.notification_noisy_notifications)
+                    enableVibration(true)
+                    enableLights(true)
+                })
 
         /**
          * Low notification importance: shows everywhere, but is not intrusive.
          */
-        val silentChannel = NotificationChannel(SILENT_NOTIFICATION_CHANNEL_ID,
+        notificationManager.createNotificationChannel(NotificationChannel(SILENT_NOTIFICATION_CHANNEL_ID,
                 context.getString(R.string.notification_silent_notifications),
                 NotificationManager.IMPORTANCE_LOW)
-        silentChannel.description = context.getString(R.string.notification_silent_notifications)
-        silentChannel.setSound(null, null)
-        silentChannel.enableLights(true)
-        notificationManager.createNotificationChannel(silentChannel)
+                .apply {
+                    description = context.getString(R.string.notification_silent_notifications)
+                    setSound(null, null)
+                    enableLights(true)
+                })
 
-        val listeningForEventChannel = NotificationChannel(LISTENING_FOR_EVENTS_NOTIFICATION_CHANNEL_ID,
+        notificationManager.createNotificationChannel(NotificationChannel(LISTENING_FOR_EVENTS_NOTIFICATION_CHANNEL_ID,
                 context.getString(R.string.notification_listening_for_events),
                 NotificationManager.IMPORTANCE_MIN)
-        listeningForEventChannel.description = context.getString(R.string.notification_listening_for_events)
-        listeningForEventChannel.setSound(null, null)
-        listeningForEventChannel.setShowBadge(false)
-        notificationManager.createNotificationChannel(listeningForEventChannel)
+                .apply {
+                    description = context.getString(R.string.notification_listening_for_events)
+                    setSound(null, null)
+                    setShowBadge(false)
+                })
 
-        val callChannel = NotificationChannel(CALL_NOTIFICATION_CHANNEL_ID,
+        notificationManager.createNotificationChannel(NotificationChannel(CALL_NOTIFICATION_CHANNEL_ID,
                 context.getString(R.string.call),
                 NotificationManager.IMPORTANCE_HIGH)
-        callChannel.description = context.getString(R.string.call)
-        callChannel.setSound(null, null)
-        callChannel.enableLights(true)
-        callChannel.lightColor = Color.GREEN
-        notificationManager.createNotificationChannel(callChannel)
+                .apply {
+                    description = context.getString(R.string.call)
+                    setSound(null, null)
+                    enableLights(true)
+                    lightColor = Color.GREEN
+                })
     }
 
     /**
@@ -180,7 +185,6 @@ object NotificationUtils {
         val pi = PendingIntent.getActivity(context, 0, i, 0)
 
         val builder = NotificationCompat.Builder(context, LISTENING_FOR_EVENTS_NOTIFICATION_CHANNEL_ID)
-                .setWhen(System.currentTimeMillis())
                 .setContentTitle(context.getString(R.string.riot_app_name))
                 .setContentText(context.getString(subTitleResId))
                 .setSmallIcon(R.drawable.logo_transparent)
@@ -232,7 +236,6 @@ object NotificationUtils {
     fun buildIncomingCallNotification(context: Context, isVideo: Boolean, roomName: String, matrixId: String, callId: String): Notification {
 
         val builder = NotificationCompat.Builder(context, CALL_NOTIFICATION_CHANNEL_ID)
-                .setWhen(System.currentTimeMillis())
                 .setContentTitle(ensureTitleNotEmpty(context, roomName))
                 .apply {
                     if (isVideo) {
@@ -287,7 +290,6 @@ object NotificationUtils {
             : Notification {
 
         val builder = NotificationCompat.Builder(context, CALL_NOTIFICATION_CHANNEL_ID)
-                .setWhen(System.currentTimeMillis())
                 .setContentTitle(ensureTitleNotEmpty(context, roomName))
                 .apply {
                     if (isVideo) {
@@ -325,21 +327,26 @@ object NotificationUtils {
         return builder.build()
     }
 
+    /**
+     * Build a notification for a Room
+     */
     fun buildMessagesListNotification(context: Context,
-                                      messageSytle: NotificationCompat.MessagingStyle,
+                                      messageStyle: NotificationCompat.MessagingStyle,
                                       roomInfo: RoomEventGroupInfo,
                                       largeIcon: Bitmap?,
+                                      lastMessageTimestamp: Long,
                                       senderDisplayNameForReplyCompat: String?): Notification? {
 
-        val accentColor = ThemeUtils.getColor(context, R.attr.colorAccent)
+        val accentColor = ContextCompat.getColor(context, R.color.notification_accent_color)
         // Build the pending intent for when the notification is clicked
         val openRoomIntent = buildOpenRoomIntent(context, roomInfo.roomId)
         val smallIcon = if (roomInfo.shouldBing) R.drawable.icon_notif_important else R.drawable.logo_transparent
 
         val channelID = if (roomInfo.shouldBing) NOISY_NOTIFICATION_CHANNEL_ID else SILENT_NOTIFICATION_CHANNEL_ID
         return NotificationCompat.Builder(context, channelID)
+                .setWhen(lastMessageTimestamp)
                 // MESSAGING_STYLE sets title and content for API 16 and above devices.
-                .setStyle(messageSytle)
+                .setStyle(messageStyle)
 
                 // A category allows groups of notifications to be ranked and filtered – per user or system settings.
                 // For example, alarm notifications should display before promo notifications, or message from known contact
@@ -354,14 +361,15 @@ object NotificationUtils {
                 // Number of new notifications for API <24 (M and below) devices.
                 .setSubText(context
                         .resources
-                        .getQuantityString(R.plurals.room_new_messages_notification, messageSytle.messages.size, messageSytle.messages.size)
+                        .getQuantityString(R.plurals.room_new_messages_notification, messageStyle.messages.size, messageStyle.messages.size)
                 )
 
                 // Auto-bundling is enabled for 4 or more notifications on API 24+ (N+)
                 // devices and all Wear devices. But we want a custom grouping, so we specify the groupID
+                // TODO Group should be current user display name
                 .setGroup(context.getString(R.string.riot_app_name))
 
-                //In order to avoid notification making sound twice (due to the summary notificaiton)
+                //In order to avoid notification making sound twice (due to the summary notification)
                 .setGroupAlertBehavior(NotificationCompat.GROUP_ALERT_SUMMARY)
 
                 .setSmallIcon(smallIcon)
@@ -385,7 +393,20 @@ object NotificationUtils {
                     }
 
                     //Add actions and notification intents
+                    // Mark room as read
+                    val markRoomReadIntent = Intent(context, NotificationBroadcastReceiver::class.java)
+                    markRoomReadIntent.action = MARK_ROOM_READ_ACTION
+                    markRoomReadIntent.data = Uri.parse("foobar://${roomInfo.roomId}")
+                    markRoomReadIntent.putExtra(NotificationBroadcastReceiver.KEY_ROOM_ID, roomInfo.roomId)
+                    val markRoomReadPendingIntent = PendingIntent.getBroadcast(context, System.currentTimeMillis().toInt(), markRoomReadIntent,
+                            PendingIntent.FLAG_UPDATE_CURRENT)
 
+                    addAction(NotificationCompat.Action(
+                            R.drawable.ic_material_done_all_white,
+                            context.getString(R.string.action_mark_room_read),
+                            markRoomReadPendingIntent))
+
+                    // Quick reply
                     if (!roomInfo.hasSmartReplyError) {
                         buildQuickReplyIntent(context, roomInfo.roomId, senderDisplayNameForReplyCompat)?.let { replyPendingIntent ->
                             val remoteInput = RemoteInput.Builder(NotificationBroadcastReceiver.KEY_TEXT_REPLY)
@@ -414,81 +435,77 @@ object NotificationUtils {
                     val pendingIntent = PendingIntent.getBroadcast(context.applicationContext,
                             System.currentTimeMillis().toInt(), intent, PendingIntent.FLAG_UPDATE_CURRENT)
                     setDeleteIntent(pendingIntent)
-                }.build()
+                }
+                .build()
     }
 
 
     fun buildSimpleEventNotification(context: Context, simpleNotifiableEvent: NotifiableEvent, largeIcon: Bitmap?, matrixId: String): Notification? {
-
-        val accentColor = ThemeUtils.getColor(context, R.attr.colorAccent)
+        val accentColor = ContextCompat.getColor(context, R.color.notification_accent_color)
         // Build the pending intent for when the notification is clicked
         val smallIcon = if (simpleNotifiableEvent.noisy) R.drawable.icon_notif_important else R.drawable.logo_transparent
 
-
         val channelID = if (simpleNotifiableEvent.noisy) NOISY_NOTIFICATION_CHANNEL_ID else SILENT_NOTIFICATION_CHANNEL_ID
-        val builder = NotificationCompat.Builder(context, channelID)
-        builder.apply {
-            setContentTitle(context.getString(R.string.riot_app_name))
-            setContentText(simpleNotifiableEvent.description)
-            //setSubText(roomInfo.roomDisplayName)
-            //setNumber(messageSytle.messages.size)
-            setGroup(context.getString(R.string.riot_app_name))
-            setGroupAlertBehavior(NotificationCompat.GROUP_ALERT_SUMMARY)
-            setSmallIcon(smallIcon)
-            color = accentColor
 
+        return NotificationCompat.Builder(context, channelID)
+                .setContentTitle(context.getString(R.string.riot_app_name))
+                .setContentText(simpleNotifiableEvent.description)
+                .setGroup(context.getString(R.string.riot_app_name))
+                .setGroupAlertBehavior(NotificationCompat.GROUP_ALERT_SUMMARY)
+                .setSmallIcon(smallIcon)
+                .setColor(accentColor)
+                .apply {
+                    if (simpleNotifiableEvent is InviteNotifiableEvent) {
+                        val roomId = simpleNotifiableEvent.roomId
+                        // offer to type a quick reject button
+                        val rejectIntent = JoinRoomActivity.getRejectRoomIntent(context, roomId, matrixId)
 
-            if (simpleNotifiableEvent is InviteNotifiableEvent) {
-                val roomId = simpleNotifiableEvent.roomId
-                // offer to type a quick reject button
-                val rejectIntent = JoinRoomActivity.getRejectRoomIntent(context, roomId, matrixId)
+                        // the action must be unique else the parameters are ignored
+                        rejectIntent.action = REJECT_ACTION
+                        rejectIntent.data = Uri.parse("foobar://$roomId&$matrixId")
+                        addAction(
+                                R.drawable.vector_notification_reject_invitation,
+                                context.getString(R.string.reject),
+                                PendingIntent.getActivity(context, System.currentTimeMillis().toInt(), rejectIntent, 0))
 
-                // the action must be unique else the parameters are ignored
-                rejectIntent.action = REJECT_ACTION
-                rejectIntent.data = Uri.parse("foobar://$roomId&$matrixId")
-                addAction(
-                        R.drawable.vector_notification_reject_invitation,
-                        context.getString(R.string.reject),
-                        PendingIntent.getActivity(context, System.currentTimeMillis().toInt(), rejectIntent, 0))
+                        // offer to type a quick accept button
+                        val joinIntent = JoinRoomActivity.getJoinRoomIntent(context, roomId, matrixId)
 
-                // offer to type a quick accept button
-                val joinIntent = JoinRoomActivity.getJoinRoomIntent(context, roomId, matrixId)
+                        // the action must be unique else the parameters are ignored
+                        joinIntent.action = JOIN_ACTION
+                        joinIntent.data = Uri.parse("foobar://$roomId&$matrixId")
+                        addAction(
+                                R.drawable.vector_notification_accept_invitation,
+                                context.getString(R.string.join),
+                                PendingIntent.getActivity(context, 0, joinIntent, 0))
 
-                // the action must be unique else the parameters are ignored
-                joinIntent.action = JOIN_ACTION + System.currentTimeMillis().toInt()
-                joinIntent.data = Uri.parse("foobar://$roomId&$matrixId")
-                addAction(
-                        R.drawable.vector_notification_accept_invitation,
-                        context.getString(R.string.join),
-                        PendingIntent.getActivity(context, 0, joinIntent, 0))
+                    } else {
+                        setAutoCancel(true)
+                    }
 
-            } else {
-                setAutoCancel(true)
-            }
+                    val contentIntent = Intent(context, VectorHomeActivity::class.java)
+                    contentIntent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP
+                    //pending intent get reused by system, this will mess up the extra params, so put unique info to avoid that
+                    contentIntent.data = Uri.parse("foobar://" + simpleNotifiableEvent.eventId)
+                    setContentIntent(PendingIntent.getActivity(context, 0, contentIntent, 0))
 
-            val contentIntent = Intent(context, VectorHomeActivity::class.java)
-            contentIntent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP
-            //pending intent get reused by system, this will mess up the extra params, so put unique info to avoid that
-            contentIntent.data = Uri.parse("foobar://" + simpleNotifiableEvent.eventId)
-            setContentIntent(PendingIntent.getActivity(context, 0, contentIntent, 0))
+                    if (largeIcon != null) {
+                        setLargeIcon(largeIcon)
+                    }
 
-            if (largeIcon != null) {
-                setLargeIcon(largeIcon)
-            }
-
-            if (simpleNotifiableEvent.noisy) {
-                //Compat
-                priority = NotificationCompat.PRIORITY_DEFAULT
-                PreferencesManager.getNotificationRingTone(context)?.let {
-                    setSound(it)
+                    if (simpleNotifiableEvent.noisy) {
+                        //Compat
+                        priority = NotificationCompat.PRIORITY_DEFAULT
+                        PreferencesManager.getNotificationRingTone(context)?.let {
+                            setSound(it)
+                        }
+                        setLights(accentColor, 500, 500)
+                    } else {
+                        priority = NotificationCompat.PRIORITY_LOW
+                    }
+                    setAutoCancel(true)
                 }
-                setLights(accentColor, 500, 500)
-            } else {
-                priority = NotificationCompat.PRIORITY_LOW
-            }
-            setAutoCancel(true)
-        }
-        return builder.build()
+                .build()
     }
 
     private fun buildOpenRoomIntent(context: Context, roomId: String): PendingIntent? {
@@ -499,10 +516,10 @@ object NotificationUtils {
         roomIntentTap.data = Uri.parse("foobar://openRoom?$roomId")
 
         // Recreate the back stack
-        val stackBuilderTap = TaskStackBuilder.create(context)
+        return TaskStackBuilder.create(context)
                 .addNextIntentWithParentStack(Intent(context, VectorHomeActivity::class.java))
                 .addNextIntent(roomIntentTap)
-        return stackBuilderTap.getPendingIntent(System.currentTimeMillis().toInt(), PendingIntent.FLAG_UPDATE_CURRENT)
+                .getPendingIntent(System.currentTimeMillis().toInt(), PendingIntent.FLAG_UPDATE_CURRENT)
     }
 
     private fun buildOpenHomePendingIntentForSummary(context: Context): PendingIntent {
@@ -524,7 +541,8 @@ object NotificationUtils {
         val intent: Intent
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             intent = Intent(context, NotificationBroadcastReceiver::class.java)
-            intent.action = "${SMART_REPLY_ACTION_PREFIX}_$roomId"
+            intent.action = SMART_REPLY_ACTION
+            intent.data = Uri.parse("foobar://$roomId")
             intent.putExtra(NotificationBroadcastReceiver.KEY_ROOM_ID, roomId)
             return PendingIntent.getBroadcast(context, System.currentTimeMillis().toInt(), intent,
                     PendingIntent.FLAG_UPDATE_CURRENT)
@@ -536,7 +554,8 @@ object NotificationUtils {
                 quickReplyIntent.putExtra(LockScreenActivity.EXTRA_SENDER_NAME, senderName ?: "")
 
                 // the action must be unique else the parameters are ignored
-                quickReplyIntent.action = QUICK_LAUNCH_ACTION + System.currentTimeMillis().toInt()
+                quickReplyIntent.action = QUICK_LAUNCH_ACTION
+                quickReplyIntent.data = Uri.parse("foobar://$roomId")
                 return PendingIntent.getActivity(context, 0, quickReplyIntent, 0)
             }
         }
@@ -544,12 +563,21 @@ object NotificationUtils {
     }
 
     //// Number of new notifications for API <24 (M and below) devices.
-    fun buildSummaryListNotification(context: Context, sytle: NotificationCompat.Style, compatSummary: String, noisy: Boolean): Notification? {
-
+    /**
+     * Build the summary notification
+     */
+    fun buildSummaryListNotification(context: Context,
+                                     style: NotificationCompat.Style,
+                                     compatSummary: String,
+                                     noisy: Boolean,
+                                     lastMessageTimestamp: Long): Notification? {
+        val accentColor = ContextCompat.getColor(context, R.color.notification_accent_color)
         val smallIcon = if (noisy) R.drawable.icon_notif_important else R.drawable.logo_transparent
 
         return NotificationCompat.Builder(context, if (noisy) NOISY_NOTIFICATION_CHANNEL_ID else SILENT_NOTIFICATION_CHANNEL_ID)
-                .setStyle(sytle) // used in compat < N, after summary is built based on child notifications
+                // used in compat < N, after summary is built based on child notifications
+                .setWhen(lastMessageTimestamp)
+                .setStyle(style)
                 .setContentTitle(context.getString(R.string.riot_app_name))
                 .setSmallIcon(smallIcon)
                 //set content text to support devices running API level < 24
@@ -557,7 +585,7 @@ object NotificationUtils {
                 .setGroup(context.getString(R.string.riot_app_name))
                 //set this notification as the summary for the group
                 .setGroupSummary(true)
-                .setColor(ThemeUtils.getColor(context, R.attr.colorAccent))
+                .setColor(accentColor)
                 .apply {
                     if (noisy) {
                         //Compat
@@ -565,6 +593,7 @@ object NotificationUtils {
                         PreferencesManager.getNotificationRingTone(context)?.let {
                             setSound(it)
                         }
+                        setLights(accentColor, 500, 500)
                     } else {
                         //compat
                         priority = NotificationCompat.PRIORITY_LOW